+++ lib/main.dart
@@ -19,7 +19,7 @@ class MyApp extends StatelessWidget {
       home: MainSplashPage(),
       debugShowCheckedModeBanner: false,
       theme: theme(),
-      initialRoute: Move.MainSplashPage,
+      initialRoute: Move.BookDetailPage,
       routes: getRouters(),
     );
   }
